# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blogs Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "as a user i can register by email and password",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "user_name",
                                    "email",
                                    "phone_number",
                                    "password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Noura"
                                    },
                                    "user_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Noura19"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Noura@example.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "0555555555"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Garage1234"
                                    },
                                    "confirm_password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Garage1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User successfully registered"
                    },
                    "404": {
                        "description": "validator error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "as a user i can login by email or username and password",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Noura"
                                    },
                                    "user_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Noura19"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Garage1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successfully"
                    },
                    "401": {
                        "description": "invalid credentials"
                    }
                }
            }
        },
        "/forgot": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "as a user i can reset my passsword",
                "operationId": "forgot",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "muhnd@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A code has been sent to your email address"
                    },
                    "404": {
                        "description": "Incorrect Email Address"
                    }
                }
            }
        },
        "/reset": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "as a user i can reset my passsword",
                "operationId": "reset",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "muhnd@example.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1234"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Garage1234"
                                    },
                                    "confirm_password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Garage1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password Reset Success"
                    },
                    "404": {
                        "description": "Token mismatch"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "as a user i can log out",
                "operationId": "logout",
                "parameters": [
                    {
                        "name": "bearerAuth",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "String"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "application/json",
                        "required": true,
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                }
            }
        },
...
    },
...
}